+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fixed issues `Delete` in Fabric file system provider
 - reworked the way how item types for the Fabric file system provider can be configured
 - minor fixes for error reporting
+- added configuration setting for `publishOnSave` when configuring ItemType Formats via `fabricStudio.itemTypeFormats`
 
 **v0.9.9**:
 - added [Workspace Role Assignments](https://learn.microsoft.com/en-us/rest/api/fabric/core/workspaces/list-workspace-role-assignments?tabs=HTTP)